refactored conversational_retrieval.py
