+++ libs/langchain/langchain/agents/agent_toolkits/conversational_retrieval.py
@@ -1,27 +1,15 @@
 from typing import Any, List, Optional
 
-from langchain.agents.agent import AgentExecutor
+from langchain.agents import AgentExecutor, OpenAIFunctionsAgent
 from langchain.agents.openai_functions_agent.agent_token_buffer_memory import (
     AgentTokenBufferMemory,
 )
-from langchain.agents.openai_functions_agent.base import OpenAIFunctionsAgent
-from langchain.chat_models.openai import ChatOpenAI
-from langchain.memory.token_buffer import ConversationTokenBufferMemory
-from langchain.prompts.chat import MessagesPlaceholder
+from langchain.chat_models import ChatOpenAI
+from langchain.memory import ConversationTokenBufferMemory
+from langchain.prompts import MessagesPlaceholder
+from langchain.schema import BaseMemory, BaseRetriever, SystemMessage
 from langchain.schema.language_model import BaseLanguageModel
-from langchain.schema.memory import BaseMemory
-from langchain.schema.messages import SystemMessage
-from langchain.tools.base import BaseTool
-
-
-def _get_default_system_message() -> SystemMessage:
-    return SystemMessage(
-        content=(
-            "Do your best to answer the questions. "
-            "Feel free to use any tools available to look up "
-            "relevant information, only if necessary"
-        )
-    )
+from langchain.tools import BaseTool, Tool
 
 
 def create_conversational_retrieval_agent(
@@ -37,18 +25,18 @@ def create_conversational_retrieval_agent(
     """A convenience method for creating a conversational retrieval agent.
 
     Args:
-        llm: The language model to use, should be ChatOpenAI
-        tools: A list of tools the agent has access to
+        llm: The language model to use, should be ChatOpenAI.
+        tools: A list of tools the agent has access to.
         remember_intermediate_steps: Whether the agent should remember intermediate
             steps or not. Intermediate steps refer to prior action/observation
-            pairs from previous questions. The benefit of remembering these is if
+            pairs from previous questions. The benefit of remembering is if
             there is relevant information in there, the agent can use it to answer
-            follow up questions. The downside is it will take up more tokens.
+            follow-up questions. The downside is it will take up more tokens.
         memory_key: The name of the memory key in the prompt.
         system_message: The system message to use. By default, a basic one will
             be used.
-        verbose: Whether or not the final AgentExecutor should be verbose or not,
-            defaults to False.
+        verbose: Whether the final AgentExecutor should be verbose or not.
+            Defaults to False.
         max_token_limit: The max number of tokens to keep around in memory.
             Defaults to 2000.
 
@@ -84,4 +72,34 @@ def create_conversational_retrieval_agent(
         verbose=verbose,
         return_intermediate_steps=remember_intermediate_steps,
         **kwargs
+    )
+
+
+def _get_default_system_message() -> SystemMessage:
+    return SystemMessage(
+        content=(
+            "Do your best to answer the questions. "
+            "Feel free to use any tools available to look up "
+            "relevant information, only if necessary"
+        )
+    )
+
+
+def create_retriever_tool(
+    retriever: BaseRetriever, name: str, description: str
+) -> Tool:
+    """Create a tool to do retrieval of documents.
+
+    Args:
+        retriever: The retriever to use.
+        name: The name for the tool. This will be passed to the language model,
+            so it should be unique and somewhat descriptive.
+        description: The description for the tool. This will be passed to the language
+            model, so should be descriptive.
+
+    Returns:
+        A Tool class.
+    """
+    return Tool(
+        name=name, description=description, func=retriever.get_relevant_documents
     )
